Added unit test for cyclic styles with palettes
